Refactor app store links
