+++ apps/store/src/components/Header/TopMenuMobile/TopMenuMobile.tsx
@@ -4,7 +4,8 @@ import { useTranslation } from 'next-i18next'
 import { useRouter } from 'next/router'
 import React, { useEffect } from 'react'
 import { AndroidIcon, AppleIcon, Button, mq, Space, theme } from 'ui'
-import { appStoreLinks } from '@/utils/appStoreLinks'
+import { getAppStoreLink } from '@/utils/appStoreLinks'
+import { useCurrentLocale } from '@/utils/l10n/useCurrentLocale'
 import {
   focusableStyles,
   MENU_BAR_HEIGHT_MOBILE,
@@ -59,6 +60,7 @@ export const TopMenuMobile = (props: TopMenuMobileProps) => {
   const { children, isOpen, setIsOpen, defaultValue } = props
   const router = useRouter()
   const { t } = useTranslation('common')
+  const { routingLocale } = useCurrentLocale()
 
   useEffect(() => {
     const closeDialog = () => setIsOpen(false)
@@ -80,7 +82,7 @@ export const TopMenuMobile = (props: TopMenuMobileProps) => {
               <div>{children}</div>
               <ButtonWrapper x={0.25}>
                 <Button
-                  href={appStoreLinks.apple}
+                  href={getAppStoreLink('apple', routingLocale).toString()}
                   target="_blank"
                   rel="noopener noreferrer"
                   variant="secondary"
@@ -90,7 +92,7 @@ export const TopMenuMobile = (props: TopMenuMobileProps) => {
                   App Store
                 </Button>
                 <Button
-                  href={appStoreLinks.google}
+                  href={getAppStoreLink('google', routingLocale).toString()}
                   target="_blank"
                   rel="noopener noreferrer"
                   variant="secondary"
